requirements.txt
numpy~=1.26.2
pandas~=2.1.3
requests~=2.31.0
seaborn~=0.13.0
pyeda~=0.29.0
PyYAML~=6.0.1